moved data fetching and side effects to custom hook
