+++ src/screens/MainScreen/MainScreen.tsx
@@ -1,56 +1,22 @@
-import React, {FC, useCallback, useState} from 'react';
-import {
-  View,
-  StyleSheet,
-  TextInput,
-  FlatList,
-  Text,
-  Image,
-  TouchableOpacity,
-  ActivityIndicator,
-} from 'react-native';
+import React, {FC} from 'react';
+import {View, TextInput, FlatList} from 'react-native';
 import {NativeStackScreenProps} from '@react-navigation/native-stack';
 
 import {RootStackParams} from '../../navigator';
-import {theme} from '../../theme/theme';
-import {getItems} from '../../api';
-import {removeHtml} from '../../helpers';
-import {useInfiniteQuery} from '@tanstack/react-query';
+import {styles} from './styles';
+import {useMainScreen} from './useMainScreen';
 
 type MainScreenProps = NativeStackScreenProps<RootStackParams, 'Home'>;
 
 export const MainScreen: FC<MainScreenProps> = ({}) => {
-  const [searchTerm, setSearchTerm] = useState('');
-
-  const {data, isLoading, isFetching, isError, fetchNextPage} =
-    useInfiniteQuery({
-      queryKey: ['items', searchTerm],
-      queryFn: ({pageParam}) => getItems(searchTerm, pageParam),
-      initialPageParam: 1,
-      getNextPageParam: (lastPage, allPages) => {
-        return allPages?.length + 1;
-      },
-    });
-
-  const isPagesNotEmpty = data?.pages?.length > 0;
-
-  const onEndReached = () => {
-    const isLastPage = data?.pages?.[data?.pages?.length - 1]?.length === 0;
-    if (!isFetching && !isLastPage) {
-      fetchNextPage();
-    }
-  };
-
-  const footerItem = useCallback(
-    () => (
-      <>
-        {isPagesNotEmpty && isFetching ? (
-          <ActivityIndicator size="small" />
-        ) : null}
-      </>
-    ),
-    [isFetching, isPagesNotEmpty],
-  );
+  const {
+    items,
+    searchTerm,
+    setSearchTerm,
+    onEndReached,
+    footerItem,
+    renderItem,
+  } = useMainScreen();
 
   const changeTextHandler = async (newValue: string) => {
     setSearchTerm(newValue);
@@ -65,28 +31,8 @@ export const MainScreen: FC<MainScreenProps> = ({}) => {
         style={styles.input}
       />
       <FlatList
-        data={data?.pages?.flatMap(page => page?.items?.materials || [])}
-        renderItem={({item}) => {
-          return (
-            <TouchableOpacity
-              style={{marginVertical: theme.small}}
-              onPress={() => alert('Yo')}>
-              <Image
-                style={{height: 250, marginVertical: theme.miniscule}}
-                source={{uri: item?.firstPreviewImage?.watermarked}}
-              />
-              <Text style={{marginVertical: theme.miniscule}}>
-                {removeHtml(item?.title ?? '')}
-              </Text>
-              <Text style={{marginVertical: theme.miniscule}}>
-                {item?.author?.details?.publicName}
-              </Text>
-              <Text style={{marginVertical: theme.miniscule}}>
-                {item?.price}
-              </Text>
-            </TouchableOpacity>
-          );
-        }}
+        data={items}
+        renderItem={renderItem}
         onEndReached={onEndReached}
         onEndReachedThreshold={0.2}
         ListFooterComponent={footerItem}
@@ -94,14 +40,3 @@ export const MainScreen: FC<MainScreenProps> = ({}) => {
     </View>
   );
 };
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    width: '100%',
-    padding: theme.medium,
-  },
-  input: {
-    height: theme.large,
-  },
-});
