+++ src/screens/MainScreen/useMainScreen.tsx
@@ -1,0 +1,69 @@
+import React, {useCallback, useState} from 'react';
+import {ActivityIndicator, Image, Text, TouchableOpacity} from 'react-native';
+import {useInfiniteQuery} from '@tanstack/react-query';
+import {getItems} from '../../api';
+import {removeHtml} from '../../helpers';
+import {theme} from '../../theme/theme';
+import {styles} from './styles';
+
+export const useMainScreen = () => {
+  const [searchTerm, setSearchTerm] = useState('');
+
+  const {data, isFetching, fetchNextPage} = useInfiniteQuery({
+    queryKey: ['items', searchTerm],
+    queryFn: ({pageParam}) => getItems(searchTerm, pageParam),
+    initialPageParam: 1,
+    getNextPageParam: (lastPage, allPages) => {
+      return allPages?.length + 1;
+    },
+  });
+
+  const isPagesNotEmpty = data?.pages?.length > 0;
+
+  const onEndReached = () => {
+    const isLastPage = data?.pages?.[data?.pages?.length - 1]?.length === 0;
+    if (!isFetching && !isLastPage) {
+      fetchNextPage();
+    }
+  };
+
+  const footerItem = useCallback(
+    () => (
+      <>
+        {isPagesNotEmpty && isFetching ? (
+          <ActivityIndicator size="small" />
+        ) : null}
+      </>
+    ),
+    [isFetching, isPagesNotEmpty],
+  );
+
+  const renderItem = useCallback(({item}) => {
+    return (
+      <TouchableOpacity
+        style={styles.renderItemContainer}
+        onPress={() => alert('Yo')}>
+        <Image
+          style={styles.renderItemImage}
+          source={{uri: item?.firstPreviewImage?.watermarked}}
+        />
+        <Text style={styles.renderItemText}>
+          {removeHtml(item?.title ?? '')}
+        </Text>
+        <Text style={styles.renderItemText}>
+          {item?.author?.details?.publicName}
+        </Text>
+        <Text style={styles.renderItemText}>{item?.price}</Text>
+      </TouchableOpacity>
+    );
+  }, []);
+
+  return {
+    items: data?.pages?.flatMap(page => page?.items?.materials || []),
+    searchTerm,
+    setSearchTerm,
+    onEndReached,
+    footerItem,
+    renderItem,
+  };
+};
